usunąłem stary skrypt. Dodałem wybieranie proposal po id. Dodałem wyjście z skryptu przez użytkownika
